Pin hugging face docs (#28106)
